requirements.txt
pdfplumber>=0.11
fastapi[standard]>=0.111
uvicorn[standard]>=0.29
python-multipart>=0.0.9
python-dotenv>=1.0
